Add main functionalitys required by lesson
I added the initial function that will populate
the container div with the required amount of
child divs. Then i made an event listener to check
for a mouseover on the child divs, changing their
background color on the event.

finally i made an event listener for the "size"
button, the listener prompts the user for a number
and then removes the old grid and replaces it
with a new one, fitting the number given. The num-
ber is for the amount of squares per side.