Add extra credit behavior
Modified divMaker function to take in the size of
an single side of the grid, making it easier to
call it on the buttons event listener that resizes
the grid.

Added more actions to event listener, these change
the color of a square to a random rgb color and
increased the opacity by 0.1 (initial value was
set to 0.1).